requirements.txt
absl-py==1.4.0
anyconfig==0.13.0
anyio==3.6.2
astunparse==1.6.3
attrs==22.2.0
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
editdistance==0.6.2
filelock==3.11.0
flake8==6.0.0
Flask==2.2.3
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
gevent==22.10.2
gevent-websocket==0.10.1
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.3
identify==2.5.22
idna==3.4
imageio==2.27.0
imgaug==0.2.8
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
kiwisolver==1.4.4
lanms-neo==1.0.2
lazy_loader==0.2
libclang==16.0.0
libcst==0.4.9
lit==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
ml-dtypes==0.0.4
moreorless==0.4.0
mpmath==1.3.0
munch==2.5.0
mypy-extensions==1.0.0
networkx==3.1
nodeenv==1.7.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
openvino==2022.3.0
opt-einsum==3.3.0
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
polygon==1.1.0
pre-commit==3.2.2
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycodestyle==2.10.0
pydocstyle==6.3.0
pyflakes==3.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
sentencepiece==0.1.96
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
stdlibs==2022.10.9
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tifffile==2023.3.21
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
trailrunner==1.4.0
transformers==4.27.2
triton==2.0.0
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
ufmt==2.1.0
urllib3==1.26.15
usort==1.0.6
validators==0.20.0
virtualenv==20.21.0
websocket-client==1.5.1
websockets==11.0.1
Werkzeug==2.2.3
wrapt==1.14.1
zope.event==4.6
zope.interface==6.0
